Create Interview_secondSTage.py
This is a code from the second stage of interview for Junior Full Stack Developer at Tenzo in London.

The main idea of this challenge is to explain why the code (names changed as they were less descriptive) does output True or False.

Comment 1: 
Variable names were changed

Comment 2: 
The code itself is not perfect, BUT that is not the point of the exercise. For example, second while loop almost never gets fired in examples provided - might be just a filler. 

Comment 3: 
I left the print statement (added by me) for better visibility of how variables change when the algorithm is performing its checks.